requirements.txt
torch
torchvision
scikit-learn
tqdm
Pillow